+++ packages/blocks/src/database-block/table/consts.ts
@@ -6,3 +6,4 @@ export const DEFAULT_COLUMN_MIN_WIDTH = 100;
 export const DEFAULT_COLUMN_TITLE_HEIGHT = 36;
 /** column title height */
 export const DEFAULT_ADD_BUTTON_WIDTH = 40;
+export const LEFT_TOOL_BAR_WIDTH = 8;
